Fix _find_element_in_node to cover all input arguments
Otherwise _translate_browsepath_to_nodeid behaves incorrect in various
situations. For example if requesting a browse path translation for
['2:X', '2:X', '2:X'] it would have returned the first object instead of
the last one as inverse references haven't been treated properly.

Now we treat the inverse flag, the reference type and its subtypes when
translating a browse path.

Additionally we're checking if the last relativepath element has a
targetname specified.

This brings the implementation in sync with
	OPC-UA Part 4: Services, 5.8.4 TranslateBrowsePathsToNodeIds